:recycle: masterInfo컴포넌트 생성 및 나눔
