+++ src/pageComponents/masters/DetailCurriculum/Curriculum/MasterInfo/index.ts
@@ -1,0 +1,1 @@
+export { default as MasterInfo } from "./MasterInfo";
